ml-module/requirements.txt
torch>=2.3.0
torchvision>=0.18.0
pillow>=10.0.0
matplotlib>=3.8.0
